# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PromptEnchanter",
    "description": "\n        PromptEnchanter is an enterprise-grade service for enhancing user prompts with AI-powered research and prompt engineering techniques.\n        \n        ## Features\n        \n        * **Prompt Engineering Styles**: Choose from various prompt engineering methodologies (BPE, BCOT, HCOT, ReAct, ToT)\n        * **AI Deep Research**: Automatic research enhancement with internet access and multi-source analysis\n        * **Batch Processing**: Process multiple prompts efficiently with parallel or sequential execution\n        * **Caching**: Intelligent caching of responses and research results\n        * **Rate Limiting**: Enterprise-grade rate limiting and security\n        * **Admin Controls**: System management and monitoring capabilities\n        \n        ## Authentication\n        \n        All endpoints require Bearer token authentication using your API key.\n        \n        ## Rate Limits\n        \n        * Standard endpoints: 100 requests/minute with burst support\n        * Batch endpoints: 25 requests/minute with reduced burst\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/users/login": {
      "post": {
        "tags": [
          "user-management"
        ],
        "summary": "User login",
        "description": "Authenticate user and create session",
        "operationId": "login_user_v1_users_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/profile": {
      "get": {
        "tags": [
          "user-management"
        ],
        "summary": "Get user profile",
        "description": "Get current user's profile information",
        "operationId": "get_user_profile_v1_users_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "user-management"
        ],
        "summary": "Update user profile",
        "description": "Update user profile information",
        "operationId": "update_user_profile_v1_users_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/support/login": {
      "post": {
        "tags": [
          "support-staff"
        ],
        "summary": "Support staff login",
        "description": "Authenticate support staff and create session",
        "operationId": "support_staff_login_v1_support_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupportStaffLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportStaffLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/support/create": {
      "post": {
        "tags": [
          "support-staff"
        ],
        "summary": "Create support staff",
        "description": "Create a new support staff member (admin only)",
        "operationId": "create_support_staff_v1_support_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportStaffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/support/profile": {
      "get": {
        "tags": [
          "support-staff"
        ],
        "summary": "Get support staff profile",
        "description": "Get current support staff profile information",
        "operationId": "get_support_staff_profile_v1_support_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportStaffInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/support/permissions": {
      "get": {
        "tags": [
          "support-staff"
        ],
        "summary": "Get permissions",
        "description": "Get current support staff permissions",
        "operationId": "get_support_staff_permissions_v1_support_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Support Staff Permissions V1 Support Permissions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/support/users": {
      "get": {
        "tags": [
          "support-staff"
        ],
        "summary": "Get users list (support staff)",
        "description": "Get paginated list of users (support staff view)",
        "operationId": "get_users_support_v1_support_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "user_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Type"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "is_verified",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Verified"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Users Support V1 Support Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/support/users/{user_id}": {
      "get": {
        "tags": [
          "support-staff"
        ],
        "summary": "Get user details (support staff)",
        "description": "Get detailed information about a specific user",
        "operationId": "get_user_details_support_v1_support_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Details Support V1 Support Users  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "support-staff"
        ],
        "summary": "Update user (support staff)",
        "description": "Update user information with support staff permissions",
        "operationId": "update_user_support_v1_support_users__user_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/support/users/{user_id}/reset-password": {
      "post": {
        "tags": [
          "support-staff"
        ],
        "summary": "Reset user password (support staff)",
        "description": "Reset user password (support level and above)",
        "operationId": "reset_user_password_support_v1_support_users__user_id__reset_password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "new_password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/support/users/{user_id}/credits": {
      "put": {
        "tags": [
          "support-staff"
        ],
        "summary": "Update user credits (support staff)",
        "description": "Update user conversation credits (support level and above)",
        "operationId": "update_user_credits_support_v1_support_users__user_id__credits_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_user_credits_support_v1_support_users__user_id__credits_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/support/users/{user_id}/plan": {
      "put": {
        "tags": [
          "support-staff"
        ],
        "summary": "Update user subscription plan (support staff)",
        "description": "Update user subscription plan (support level and above)",
        "operationId": "update_user_plan_support_v1_support_users__user_id__plan_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "subscription_plan",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subscription Plan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLoginRequest"
      },
      "UserLoginResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          },
          "session": {
            "$ref": "#/components/schemas/SessionInfo"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "user",
          "session"
        ],
        "title": "UserLoginResponse"
      },
      "UserProfile": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "about_me": {
            "type": "string",
            "title": "About Me"
          },
          "hobbies": {
            "type": "string",
            "title": "Hobbies"
          },
          "user_type": {
            "type": "string",
            "title": "User Type"
          },
          "time_created": {
            "type": "string",
            "format": "date-time",
            "title": "Time Created"
          },
          "subscription_plan": {
            "type": "string",
            "title": "Subscription Plan"
          },
          "credits": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Credits"
          },
          "limits": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Limits"
          },
          "access_rtype": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Access Rtype"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "additional_notes": {
            "type": "string",
            "title": "Additional Notes"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login"
          },
          "last_activity": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Activity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "name",
          "email",
          "about_me",
          "hobbies",
          "user_type",
          "time_created",
          "subscription_plan",
          "credits",
          "limits",
          "access_rtype",
          "level",
          "additional_notes",
          "is_verified",
          "last_login",
          "last_activity"
        ],
        "title": "UserProfile"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}